tutorials/griptape/chatbots/pyproject.toml
[tool.poetry]
name = "chatbots"
version = "0.1.0"
description = ""
authors = ["Alan Yu <inv.alanyu2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
griptape = "^0.17.1"
python-dotenv = "^1.0.0"
rich = "^13.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
